S4-0: /onboarding como router — wizard first-run + Configurar cuenta admin
